Replace C-style casts with static_cast across all source files
Replaces (float), (int), (GLfloat), (GLint), (char*) casts with their
static_cast equivalents. Font.cpp char-array indexing uses
static_cast<unsigned char> to avoid sign-extension on signed char.
Redundant outer casts on already-float expressions removed.

+++ src/framework/Utils.cpp
@@ -25,7 +25,7 @@ namespace Hunchback::Framework {
 
 int roundi(double x)
 {
-	return (int)(x + 0.5);
+	return static_cast<int>(x + 0.5);
 }
 
 int nextpoweroftwo(int x)
@@ -46,11 +46,11 @@ int power_of_two(int input)
 }
 
 int xformX(int x) {
-	return (int)( (float((float)(x) / (float)(SCREEN_DIM_W)) * MediaManager::getScreenWidth()) );
+	return static_cast<int>(static_cast<float>(x) / static_cast<float>(SCREEN_DIM_W) * MediaManager::getScreenWidth());
 }
 
 int xformY(int y) {
-	return (int)( (float((float)(y) / (float)(SCREEN_DIM_H)) * MediaManager::getScreenHeight()) );
+	return static_cast<int>(static_cast<float>(y) / static_cast<float>(SCREEN_DIM_H) * MediaManager::getScreenHeight());
 }
 
 } // namespace Hunchback::Framework
